+++ plugins/LinuxVST/src/PowerSag2/PowerSag2Proc.cpp
@@ -20,12 +20,12 @@ void PowerSag2::processReplacing(float **inputs, float **outputs, VstInt32 sampl
     
     while (--sampleFrames >= 0)
     {
-		long double inputSampleL = *in1;
-		long double inputSampleR = *in2;
+		double inputSampleL = *in1;
+		double inputSampleR = *in2;
 		if (fabs(inputSampleL)<1.18e-37) inputSampleL = fpd * 1.18e-37;
 		if (fabs(inputSampleR)<1.18e-37) inputSampleR = fpd * 1.18e-37;
-		long double drySampleL = inputSampleL;
-		long double drySampleR = inputSampleR;
+		double drySampleL = inputSampleL;
+		double drySampleR = inputSampleR;
 		
 		if (gcount < 0 || gcount > 16384) {gcount = 16384;}		
 		dL[gcount] = fabs(inputSampleL);
@@ -89,12 +89,12 @@ void PowerSag2::processDoubleReplacing(double **inputs, double **outputs, VstInt
 	
     while (--sampleFrames >= 0)
     {
-		long double inputSampleL = *in1;
-		long double inputSampleR = *in2;
+		double inputSampleL = *in1;
+		double inputSampleR = *in2;
 		if (fabs(inputSampleL)<1.18e-43) inputSampleL = fpd * 1.18e-43;
 		if (fabs(inputSampleR)<1.18e-43) inputSampleR = fpd * 1.18e-43;
-		long double drySampleL = inputSampleL;
-		long double drySampleR = inputSampleR;
+		double drySampleL = inputSampleL;
+		double drySampleR = inputSampleR;
 		
 		if (gcount < 0 || gcount > 16384) {gcount = 16384;}		
 		dL[gcount] = fabs(inputSampleL);
